requirements.txt
streamlit
requests
numpy
qrcode
Pillow